+++ src/hw/pvr/tr.c
@@ -988,7 +988,7 @@ void tr_parse_context(struct tr *tr, const struct tile_ctx *ctx,
 
     /* track info about the parse state for tracer debugging */
     struct render_param *rp = &rc->params[rc->num_params++];
-    rp->offset = data - ctx->params;
+    rp->offset = (int)(data - ctx->params);
     rp->list_type = tr->list_type;
     rp->vertex_type = tr->list_type;
     rp->last_surf = rc->num_surfs - 1;
